flip bits

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -1,14 +1,14 @@
 #include "holberton.h"
 
 /**
- * flip_bits - return the number of bits you need to flip number to other number
+ * flip_bits - return the number of bits you need to flip number
  * @n: a given number to change
  * @m: number expected after the change
  * Return: number of bits
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	int bits = 0;
+	unsigned long int bits = 0;
 	int xor;
 
 	for (xor = n ^ m; xor != 0; xor = xor >> 1)
